# perf(SendersRecovery): re-use Secp256K1 context for >2x speedup and add benches