--- a/crates/stages/Cargo.toml
+++ b/crates/stages/Cargo.toml
@@ -49,6 +49,7 @@ proptest = { version = "1.0", optional = true }
 
 [dev-dependencies]
 # reth
+reth-primitives = { path = "../primitives", features = ["arbitrary"]}
 reth-db = { path = "../storage/db", features = ["test-utils", "mdbx"] }
 reth-interfaces = { path = "../interfaces", features = ["test-utils"] }
 reth-downloaders = { path = "../net/downloaders" }
@@ -59,9 +60,10 @@ assert_matches = "1.5.0"
 rand = "0.8.5"
 paste = "1.0"
 
-# arbitrary utils
-arbitrary = { version = "1.1.7", features = ["derive"] }
+# Stage benchmarks
+criterion = { version = "0.4.0", features = ["async_futures"] }
 proptest = { version = "1.0" }
+arbitrary = { version = "1.1.7", features = ["derive"] }
 
 # trie
 reth-staged-sync = { path = "../staged-sync" }
@@ -70,3 +72,10 @@ triehash = "0.8"
 [features]
 default = ["serde"]
 serde = ["dep:serde"]
+test-utils = []
+
+
+[[bench]]
+name = "criterion"
+harness = false
+required-features = ["test-utils"]