--- a/crates/stages/benches/README.md
+++ b/crates/stages/benches/README.md
@@ -1,0 +1,8 @@
+# Stage Benchmarks
+
+Test vectors are automatically generated if they cannot be found.
+
+## Usage
+```
+cargo bench --package reth-stages --bench criterion --features test-utils
+```