--- a/crates/stages/src/lib.rs
+++ b/crates/stages/src/lib.rs
@@ -57,8 +57,9 @@ mod stage;
 mod trie;
 mod util;
 
-#[cfg(test)]
-mod test_utils;
+#[allow(missing_docs)]
+#[cfg(any(test, feature = "test-utils"))]
+pub mod test_utils;
 
 /// A re-export of common structs and traits.
 pub mod prelude;