--- a/crates/stages/src/stages/tx_lookup.rs
+++ b/crates/stages/src/stages/tx_lookup.rs
@@ -17,7 +17,7 @@ const TRANSACTION_LOOKUP: StageId = StageId("TransactionLookup");
 /// This stage walks over the bodies table, and sets the transaction hash of each transaction in a
 /// block to the corresponding `TransitionId` at each block. This is written to the
 /// [`tables::TxHashNumber`] This is used for looking up changesets via the transaction hash.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct TransactionLookupStage {
     /// The number of table entries to commit at once
     commit_threshold: u64,