--- a/crates/stages/src/test_utils/mod.rs
+++ b/crates/stages/src/test_utils/mod.rs
@@ -1,3 +1,4 @@
+#![allow(unused)]
 use crate::StageId;
 
 mod macros;