# Adding of Benchmarks

Some overlap with #1130, I decided to go with a different test scaffolding method. 
* Instead of using raw files, let's always try to create a test database at a specific path, and have it be re-used as much as possible. In the case of senders recovery the data also needs to have valid signatures, so we re-use the methods from `reth_interfaces::test_utils::generators` which can give us a valid chain of blocks (vs using `arbitrary` like in #1130).
* I also run it with a few diff params, to get a sense of how does the batch size affect perf. Blocks have ~150 txs lately, so I think with 10k commit threshold and 250k batch size should be good.

Given the same test vectors can be reused to benchmark the performance of the txlookup stage, a bench is also added for `TransactionLookup` stage.

# Perf improvements in recovery

Let's get in the habit of sharing criterion bench before/after and explain how that worked.

We were previously re-creating the context for secp via `Secp256k1::new()`, when instead we could re-use it. Given this was happening in a giga-hot loop (there's 1.8B transactions!) this ended up taking a significant amount of time. Improves perf ~3x.

Main w/ 100K vs 1k batch

```
# use the name of the measurement to only run these (this will skip tx lookup)
cargo bench --features=test-utils -- Senders

Benchmarking Stages/SendersRecovery-batch-10000: Warming up for 2.0000 s

Stages/SendersRecovery-batch-1000
                        time:   [15.967 s 16.869 s 17.949 s]
                        change: [+123.59% +136.91% +153.67%] (p = 0.00 < 0.05)
                        Performance has regressed.

Benchmarking Stages/SendersRecovery-batch-100000: Warming up for 2.0000 s
Warning: Unable to complete 10 samples in 2.0s. You may wish to increase target time to 148.9s.
Stages/SendersRecovery-batch-100000
                        time:   [14.352 s 15.115 s 16.126 s]
                        change: [+128.10% +139.93% +155.75%] (p = 0.00 < 0.05)
                        Performance has regressed.

```

So it makes sense to bump the batch a bunch, I've pushed to 250K here (incl in the bench).

And with the fix (`cargo bench --features=test-utils -- SendersRecovery-batch-100000` to only run the 100k batch)

```
Stages/SendersRecovery-batch-100000
                        time:   [6.4679 s 6.6515 s 6.8637 s]
                        change: [-59.031% -55.994% -53.332%] (p = 0.00 < 0.05)
                        Performance has improved.
```

cc @joshieDo 

As a follow-on optimization here, we could have a LRU cache of PubKey->Address that gets populated for the most popular 100K addresses, to avoid hashing for them.